Feature - assert_match with ignore_keys
Enable setting ignore_keys argument on snapshot.assert_match()
ignore_keys is a list or tuple of keys whose values should be ignored when running assert_match. The values of the ignore_keys are set to None, this way we still assert that the keys are present but ignore its value. Dicts, lists and tuples are traversed recursively to ignore any occurrence of the key for all dicts on any level.