Handle SortedDict with keys other than strings (#111)
`SortedDict` only allowed named arguments as parameters in the form of `**kwargs`, hence limiting the creation of `SortedDict`s with keys other than strings (fixes #65). This PR replaces `**kwargs` as the main argument to build a `SortedDict` and uses instead a regular dict `values` that can contain anything a regular Python dict could.

The only significant change is that to instantiate a new `SortedDict` now, you don't need to pass in the named arguments, `SortedDict(**value)`, but just the variable, `SortedDict(value)`.

Tests have been added for the `SortedDict` class.